requirements.txt
Django>=5.2,<5.3
PyMySQL>=1.1,<2.0
python-dotenv>=1.0,<2.0
PyJWT>=2.10,<3.0
Pillow>=11.0,<12.0
